+++ demos/LuaGLWidget.hpp
@@ -53,7 +53,7 @@ private:
 		return heights.get(x, y);
 	}
 protected slots:
-	void tick(const float& elapsed);
+	void tick(const double& elapsed);
 protected:
 	void render();
 	void initializeGL();
